feat(condo): DOMA-11362 send ticket created and ticket comment created notifications in related organizations
